Finish migration of fakes.dart, and combine with remove migrated_fakes.dart
