+++ dwds/test/debugger_test.dart
@@ -21,7 +21,7 @@ import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart'
 import 'fixtures/context.dart';
 import 'fixtures/debugger_data.dart';
 import 'fixtures/fakes.dart';
-import 'fixtures/migrated_fakes.dart';
+import 'fixtures/fakes.dart';
 
 final context = TestContext();
 AppInspector inspector;
